fix: sudo is not needed for command : ansible-galaxy install geerlingguy.mysql
as a matter of fact that sudo command is a way to become root privileges, using sudo will cause deployment of roles from ansible-galaxy to current user profile - that is root but not vagrant as expected. So we need to remove sudo and simple run install command for role. Also kernel and console log of VM contain a lot of sensitive data and should be added to .gitignore